requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.29
requests==2.31.0
passlib==1.7.4
python-dotenv==1.0.1
